FIX-#2054: Moved non-dependent on modin.DataFrame utils to modin/utils.py  (#2055)
