+++ modin/data_management/functions/groupby_function.py
@@ -14,7 +14,7 @@
 import pandas
 
 from .mapreducefunction import MapReduceFunction
-from modin.pandas.utils import try_cast_to_pandas
+from modin.utils import try_cast_to_pandas
 
 
 class GroupbyReduceFunction(MapReduceFunction):
