+++ modin/pandas/series.py
@@ -24,10 +24,10 @@ from pandas.core.dtypes.common import (
 import sys
 import warnings
 
+from modin.utils import _inherit_docstrings, to_pandas
 from .base import BasePandasDataset
 from .iterator import PartitionIterator
-from .utils import _inherit_docstrings
-from .utils import from_pandas, to_pandas, is_scalar
+from .utils import from_pandas, is_scalar
 
 if sys.version_info[0] == 3 and sys.version_info[1] >= 7:
     # Python >= 3.7
